\docType{data}
\name{data2}
\alias{count2}
\alias{coord2}

\title{Example Data of Human Breast Cancer}
\usage{
data(data2)
}

\description{
\describe{
\item{count2}{A numeric matrix of gene expression counts for 14,798 genes and 251 spots.} \item{coord2}{A numeric matrix of two-dimensioanl spatial coordinates for 251 spots.}
}
}







